requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
base58==2.1.1
bitarray==2.7.6
blocksmith==1.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
cytoolz==0.12.1
ecdsa==0.18.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parsimonious==0.8.1
protobuf==3.19.5
psutil==5.9.2
py-solc-x==1.1.1
pycryptodome==3.18.0
Pygments==2.15.1
pygments-lexer-solidity==0.7.0
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt6==6.4.0
PyQt6-QScintilla==2.13.3
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2023.3
pywin32==306
PyYAML==6.0
QScintilla==2.13.3
referencing==0.29.1
requests==2.31.0
rlp==2.0.1
rpds-py==0.8.8
semantic-version==2.10.0
six==1.16.0
toolz==0.12.0
urllib3==2.0.3
varint==1.0.2
w3storage==0.0.1
web3==5.31.1
websockets==9.1
yarl==1.9.2
